fix 3227 and adds unittest

+++ zds/forum/commons.py
@@ -150,6 +150,9 @@ class PostEditMixin(object):
         But, if there is only one post in the topic, we mark the topic unread but we don't create a notification.
         """
         topic_read = TopicRead.objects.filter(topic=post.topic, user=user).first()
+        # issue 3227 proves that you can have post.position==1 AND topic_read to None
+        # it can happen whether on double click (the event "mark as not read" is therefore sent twice)
+        # or if you have too tabs in your browser.
         if topic_read is None and post.position > 1:
             unread = Post.objects.filter(topic=post.topic, position=(post.position - 1)).first()
             topic_read = TopicRead(post=unread, topic=unread.topic, user=user)
@@ -159,7 +162,7 @@ class PostEditMixin(object):
                 unread = Post.objects.filter(topic=post.topic, position=(post.position - 1)).first()
                 topic_read.post = unread
                 topic_read.save()
-            else:
+            elif topic_read:
                 topic_read.delete()
 
         signals.answer_unread.send(sender=post.topic.__class__, instance=post, user=user)
